fifth step

+++ database.sql
@@ -10,7 +10,7 @@ DROP TABLE IF EXISTS url_checks;
 CREATE TABLE url_checks (
   id bigint PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
   url_id bigint REFERENCES urls (id),
-  status_code bigint,
+  status_code int,
   h1 text,
   title text,
   description text,
